+++ On2Cook Cloud/src/data-store.js
@@ -661,6 +661,7 @@ function createUiState() {
     },
     demoAuthBypass: false,
     activeModal: null,
+    orderNotice: null,
     toast: "",
     toastTone: "info"
   };